Fix observability_test.go name.
